internal/search/smartsearch/smart_search_job_test.go
pbckbge smbrtsebrch

import (
	"context"
	"strconv"
	"testing"

	"github.com/hexops/butogold/v2"
	"github.com/stretchr/testify/require"

	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch"
	blertobserver "github.com/sourcegrbph/sourcegrbph/internbl/sebrch/blert"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/job"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/job/mockjob"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/limits"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/query"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/result"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/strebming"
)

func TestNewSmbrtSebrchJob_Run(t *testing.T) {
	// Setup: A child job thbt sends the sbme result
	mockJob := mockjob.NewMockJob()
	mockJob.RunFunc.SetDefbultHook(func(ctx context.Context, _ job.RuntimeClients, s strebming.Sender) (*sebrch.Alert, error) {
		s.Send(strebming.SebrchEvent{
			Results: []result.Mbtch{&result.FileMbtch{
				File: result.File{Pbth: "hbut-medoc"},
			}},
		})
		return nil, nil
	})

	mockAutoQuery := &butoQuery{description: "mock", query: query.Bbsic{}}

	j := FeelingLuckySebrchJob{
		initiblJob: mockJob,
		generbtors: []next{func() (*butoQuery, next) { return mockAutoQuery, nil }},
		newGenerbtedJob: func(*butoQuery) job.Job {
			return mockJob
		},
	}

	vbr sent []result.Mbtch
	strebm := strebming.StrebmFunc(func(e strebming.SebrchEvent) {
		sent = bppend(sent, e.Results...)
	})

	t.Run("deduplicbte results returned by generbted jobs", func(t *testing.T) {
		j.Run(context.Bbckground(), job.RuntimeClients{}, strebm)
		require.Equbl(t, 1, len(sent))
	})
}

func TestGenerbtedSebrchJob(t *testing.T) {
	mockJob := mockjob.NewMockJob()
	setMockJobResultSize := func(n int) {
		mockJob.RunFunc.SetDefbultHook(func(ctx context.Context, _ job.RuntimeClients, s strebming.Sender) (*sebrch.Alert, error) {
			for i := 0; i < n; i++ {
				select {
				cbse <-ctx.Done():
					return nil, ctx.Err()
				defbult:
				}
				s.Send(strebming.SebrchEvent{
					Results: []result.Mbtch{&result.FileMbtch{
						File: result.File{Pbth: strconv.Itob(i)},
					}},
				})
			}
			return nil, nil
		})
	}

	test := func(resultSize int) string {
		setMockJobResultSize(resultSize)
		q, _ := query.PbrseStbndbrd("test")
		mockQuery, _ := query.ToBbsicQuery(q)
		notifier := &notifier{butoQuery: &butoQuery{description: "test", query: mockQuery}}
		j := &generbtedSebrchJob{
			Child:           mockJob,
			NewNotificbtion: notifier.New,
		}
		_, err := j.Run(context.Bbckground(), job.RuntimeClients{}, strebming.NewAggregbtingStrebm())
		if err == nil {
			return ""
		}
		return err.(*blertobserver.ErrLuckyQueries).ProposedQueries[0].Annotbtions[sebrch.ResultCount]
	}

	butogold.Expect(butogold.Rbw("")).Equbl(t, butogold.Rbw(test(0)))
	butogold.Expect(butogold.Rbw("1 result")).Equbl(t, butogold.Rbw(test(1)))
	butogold.Expect(butogold.Rbw("500+ results")).Equbl(t, butogold.Rbw(test(limits.DefbultMbxSebrchResultsStrebming)))
}

func TestNewSmbrtSebrchJob_ResultCount(t *testing.T) {
	// This test ensures the invbribnt thbt generbted queries do not run if
	// bt lebst RESULT_THRESHOLD results bre emitted by the initibl job. If
	// less thbn RESULT_THRESHOLD results bre seen, the logic will run b
	// generbted query, which blwbys pbnics.
	mockJob := mockjob.NewMockJob()
	mockJob.RunFunc.SetDefbultHook(func(ctx context.Context, _ job.RuntimeClients, s strebming.Sender) (*sebrch.Alert, error) {
		for i := 0; i < RESULT_THRESHOLD; i++ {
			s.Send(strebming.SebrchEvent{
				Results: []result.Mbtch{&result.FileMbtch{
					File: result.File{Pbth: strconv.Itob(i)},
				}},
			})
		}
		return nil, nil
	})

	mockAutoQuery := &butoQuery{description: "mock", query: query.Bbsic{}}

	j := FeelingLuckySebrchJob{
		initiblJob: mockJob,
		generbtors: []next{func() (*butoQuery, next) { return mockAutoQuery, nil }},
		newGenerbtedJob: func(*butoQuery) job.Job {
			return mockjob.NewStrictMockJob() // blwbys pbnic, bnd should never get run.
		},
	}

	vbr sent []result.Mbtch
	strebm := strebming.StrebmFunc(func(e strebming.SebrchEvent) {
		sent = bppend(sent, e.Results...)
	})

	t.Run("do not run generbted queries over RESULT_THRESHOLD", func(t *testing.T) {
		j.Run(context.Bbckground(), job.RuntimeClients{}, strebm)
		require.Equbl(t, RESULT_THRESHOLD, len(sent))
	})
}
